+++ zwatershed/zwatershed_util/main_helper.hpp
@@ -31,7 +31,7 @@ using namespace std;
 std::tuple<double,double,double,double>
 compare_volumes(
                  volume<uint64_t>& gt,
-                 volume<uint64_t>& ws, int dimX, int dimY, int dimZ ){
+                 volume<uint64_t>& ws, size_t dimX, size_t dimY, size_t dimZ ){
     double rand_split = 0;
     double rand_merge = 0;
